:sparkles: [#99] Make AuthenticationRequestView more general purpose
Ported the refactor from Open Forms.

The library now ships an OIDCInit view that can handle multiple config
classes (and in the future: instances of a single model), which all
have tailored behaviours. The default AuthenticationRequestView is
now 'just a flavour' of this generic behaviour.

We track some additional state and perform additional redirect URL
validation for those situations where stricter behaviour of the
OIDC flow is desired.