:sparkles: [#99] Implement utilities to extract config_class context from request/session
Since the OIDC init view now uses composition to specify which OIDC
config class should be used, we also need to be able to extract this
information again to make sure the same config class is used during
the callback flow.

The upstream library clears the state during the callback phase, so we
must make sure the config information is extracted *before* this is
done. For that reason, we extract the info from the session and store
it as a temporary attribute on the request object to 'smuggle' it to
the authentication backend.