:sparkles: [#99] Made the authentication backend flexible in terms of configuration to use
The authentication backend is now general-purpose - you don't need to
subclass it anymore and do weird tricks to make it use a different
OIDC Configuration subclass.

The relevant config_class to use is extracted from the user session,
tied to a particular OIDC state created by the init flow. This config
class is then used to look up the relevant settings to obtain the
necessary tokens and grab the user information (either from the ID
token directly or from an user-info endpoint).

The class works out of the box with our concrete `OpenIDConnectConfig`
model. Some 'settings' are not present on the `OpenIDConnectConfigBase\'
model - those are retrieved from the django settings if provided, or
fall back to a sensible default.

Users of this library can bring their own configuration models and
implement the relevant model fields/properties to populate the
settings, or simply specify them as django settings if they're fixed
for the project.