[#136] Ensure lookup_config uses session as fallback location
The lookup_config helper is now mainly used to fetch the persisted
config directly from the user object on the request. However, there
are situations, notably in the upcoming change to SessionRefresh,
where the backend config needs to be dynamically constructed from
the session in an out-of-band request where the config has not been
persisted on the user object.

This commit adds a fallback lookup to the lookup_config helper to
try and fetch the config from the session if it's not available on
the user object.