[#136] Dynamically lookup config per request in SessionRefresh
The SessionRefresh middleware was hardcoded to always use the
OpenIDConnectConfig config object, regardless of the backend
that was actually used to authenticate. This would lead to
issues in production, as users would authenticate with a
_another_ backend (say, DigiD OIDC), but would, upon hitting
the middleware's timeout, always be redirected to the config
based on OpenIDConnectConfig singleton, typically with another
backend's state, ID token hint, and so forth, leading to a
variety of errors.

This commit uses the persisted config information on the session
object to dynamically select the appropriate config object.

BREAKIN CHANGE: The BaseSessionRefreshMiddleware has been removed
as the single middleware will now handle all configured backends.