:bug: [#120] Fix retrieval of optional endpoints
Not every OpenID discovery endpoint exposes values for all
possible endpoints, in particular the not-required ones.

The current implementation fell back to use .get, which
returns None, which can't be saved in the database because
it's a string-based field that isn't nullable.

Fixed by skipping over missing/empty values, which falls back
to the model field default instead.

Additionally, string-concatenation of URL-parts is naive
and broken - efforts need to be done to guarantee necessary
slashes are added. When joining base URLs and relative
paths, one should always use urllib.parse.urljoin to
properly cover this.